requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
asgiref==3.8.1
    # via django
dj-database-url==2.0.0
    # via -r requirements.in
django==4.2
    # via
    #   -r requirements.in
    #   dj-database-url
sqlparse==0.5.0
    # via django
typing-extensions==4.12.0
    # via
    #   asgiref
    #   dj-database-url
psycopg2-binary==2.9.9
dj-database-url==2.0.0
whitenoise==6.4.0
gunicorn==20.1.0
uvicorn==0.29.0
PyYAML==6.0.1
requests==2.32.3